requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
Django==5.1.2
exceptiongroup==1.2.2
greenlet==3.0.3
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
playwright==1.47.0
pyee==11.1.1
pyppeteer==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.20.2
